Add FieldsSeq
